Updated setting text + renamed toggle file
