+++ src/renderer/pages/settings/appearance_settings/appearance_settings.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { LanguageSelector } from "./language_selector/language_selector";
-import { RankDisplayToggle } from "./rank_display_toggle/RankDisplayToggle";
+import { RankDisplayToggle } from "./rank_display_toggle/rank_display_toggle";
 
 export const AppearanceSettings = React.memo(() => {
   return (
