+++ src/renderer/pages/settings/appearance_settings/rank_display_toggle/rank_display_toggle.messages.ts
@@ -1,4 +1,4 @@
 export const RankDisplayToggleMessages = {
-  rankDisplay: () => "Rank display",
-  rankDisplayDescription: () => "Enable the current rank display on the home page.",
+  rankDisplay: () => "Show Ranking on Home Page",
+  rankDisplayDescription: () => "If enabled, your current rank and rating will be displayed on the Home page.",
 };
