+++ src/renderer/pages/settings/appearance_settings/rank_display_toggle/rank_display_toggle.tsx
